Documentation updates and support for reverse playback
Forgot to update the documentation accordingly with the prior push. More importantly, however, it is now possible to use reverse-playback variations of samples without needing to import multiple versions of the same sample. Make sure to write "REV" in the name for one of the sfz multisample variations to use this potential, as now written in the readme.