Major updates and bugfix
More updates. Most notable changes:

1) fixed encoding of samples without loop points in particular

2) fixed importing of 32-bit audio.

3) most importantly, this commit verifies working patches for the JD-990, having acquired one the other day to test with at last. This means that it should now be feasible to make more JD-exclusive patches for the cards a la SR-JV80-04.

It should also be noted: some of the revisions to the 990 patch importing subscript are out of necessity for a quirk with MIDIQuest: while it can export SysEx like what I used for testing, the offset of the SysEx headers in the third tone are incorrect, so this implements a workaround for any other MIDIQuest users hoping to test with those files.