Syntax fix and experimental frequency compensation
Commit that fixes the last one failing to run properly: it turned out an argument for one of the crucial functions was forgotten about.

That in mind, equally important: this update includes an experimental approach to making imported samples less muddy sounding a la the original data. Make sure the exponents stay under 10 for testing in the hardware, though, and the original importing mode is still enabled by default to make it easier to determine good loop points as well.