Sample importing improved
Prior commits of this project actually only used 7 bits worth of delta samples when importing. In retrospect, I still feel it was a good starting point given the most significant bit needs to keep being switched when subtracting.

That said, good news and bad news.

Good news: all eight bits should now hopefully import cleanly now, including in the experimental Bright Mode. A certain function increases the delta amount for the coefficient calculations VERY slightly to avoid some unusual cases of DC drift, but otherwise, most mono WAV samples should work, and even with that change, the sonic difference after decompression and recompression with SR-JV80-15 test samples is negligible even to a spectrogram.

Bad news: Also discovered certain samples do not cooperate with the main algorithm I recreated to address the highest frequencies with Bright Mode on, so as it is now, three attempts are made. First is the proper high frequency compensating. If that gets too loud with a given sample, then an in-between approach is used, and if that doesn't work either, then it imports as if it was disabled as a last resort.

Regardless, this should now be fully workable for many purposes, and now that I have a working JD-990, I can verify custom patch banks meant for it will load!