Experimental support for more SFZ opcodes
It should now be feasible to do without the SMPL data chunk in the WAV files imported entirely.

Currently the sample metadata is dependent on the sample's first occurrence in the multisample SFZ files, and only if no valid SMPL chunk is found in the WAV file, partially for compatibility with previous versions of these scripts, but also to prevent too much complexity with samples.