Duplicate sample fixes and variable sample rates
Implemented fixes for importing duplicate samples if they're a sample rate other than the ROM's, which is no longer fixed to 32000 in these scripts in light of the JD-990. Also prevented unnecessary risk of DC offset with reimporting official samples by appending 0x00 instead of 0xff to the end of delta samples.