Minor fix
Addressed a ROM recreation issue in which offset sample lengths would result in samples being reimported in a different order. Thankfully, this does not seem to affect the imported samples in any way.

On another note: unimplemented, but it appears that newer versions of Python will throw out warnings due to their changing the Regex escape format. Thus, if you have any issues on newer versions of Python, try editing the scripts by making two forwards slashes in the regex methods of concern instead of one. Check your command line output to find which methods specifically if you're not sure.